Add files via upload
The Date: Date: May 22, 2025 – 6:00 AM (EST) in this document and all documents created by Spencer Southern or any representative or employee is meant to establish and not to limit the prior art dates of when any IP was truly created. These types of declarations are meant to prove continuous work and to continually establish ownership. If you are unaware of the meaning of any documents created by Spencer Southern or any one or thing he says then it is your responsibility to ask and not assume. You are not allowed to make assumptions of Spencer Sothern's' work. ASK!

Intellectual Property Notice

All concepts, systems, writings, designs, and related materials created by Spencer Southern are the exclusive intellectual property of Spencer Southern / Southern Star Pro. Studios LLC.
They are protected under applicable copyright, trade‑secret, and intellectual‑property laws in the United States and internationally.
No portion of this material may be copied, reproduced, distributed, or used in any form without prior written authorization.
All rights reserved.